requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.11.0
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
frozenlist==1.8.0
git-filter-repo==2.47.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
inflection==0.5.1
instaloader==4.15
magic-filter==1.0.12
multidict==6.7.0
packaging==25.0
pillow==12.0.0
propcache==0.4.1
pydantic==2.11.10
pydantic_core==2.33.2
PyJWT==2.10.1
python-telegram-bot==21.0.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
sniffio==1.3.1
sqlparse==0.5.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
yarl==1.22.0
yt-dlp==2025.11.12
